# backend/posproject/src/main/resources/OpenAPI.yaml
openapi: 3.0.3
info:
  title: "Spring Boot REST Academia API"
  version: "1.0.0"
  description: "API for Teachers, Students and Subjects."
servers:
  - url: http://localhost:8080/api/academia
paths:
  /students:
    get:
      summary: Get all students
      security:
        - bearerAuth : []
      responses:
        '200':
          description: List of all students
        '204':
          description: No students recorded
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    post:
      summary: Create a student
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentAddDTO'
      responses:
        '201':
          description: Student Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Student already exists with these details
        '415':
          description: Unsupported Media Type
        '422':
          description: Student data does not correspond with formatting rules
  /students/{id}:
    get:
      summary: Get a student by ID
      security:
        - bearerAuth : []
      responses:
        '200':
          description: Get a student by ID
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Student Not Found

    put:
      summary: Update a student by ID
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentAddDTO'
      responses:
        '204':
          description: Successfully updated student
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Student Not Found
        '415':
          description: Unsupported Media Type
        '422':
          description: Student data does not correspond with formatting rules

    delete:
      summary: Delete a student by ID
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Successfully deleted student
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Student Not Found
  /{id}/subjects:
    get:
      summary: Get a student's subjects
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of student's subjects
        '204':
          description: Student has no subjects
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /subjects/{id}/students:
    get:
      summary: "Get students of a subject"
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: List of students enrolled in course
        '204':
          description: No students enrolled in this course
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Subject Not Found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentAddDTO:
      type: object
      properties:
        surname:
          type: string
        name:
          type: integer
        email:
          type: string
        group:
          type: string
        studyYear:
          type: integer
        studyCycle:
          type: string
        joinedSubjects:
          type: array
          items:
            type: string
...